Fix many float conversion warnings in the scripts

+++ src/scripts/Outland/HellfireCitadel/shattered_halls/boss_warchief_kargath_bladefist.cpp
@@ -11,9 +11,9 @@
 #define MOB_SHARPSHOOTER_GUARD          17622
 #define MOB_REAVER_GUARD                17623
 
-float AssassEntrance[3] = {275.136,-84.29,2.3}; // y +-8
-float AssassExit[3] = {184.233,-84.29,2.3}; // y +-8
-float AddsEntrance[3] = {306.036,-84.29,1.93};
+float AssassEntrance[3] = {275.136f,-84.29f,2.3f}; // y +-8
+float AssassExit[3] = {184.233f,-84.29f,2.3f}; // y +-8
+float AddsEntrance[3] = {306.036f,-84.29f,1.93f};
 
 #define SOUND_AGGRO1                    10323
 #define SAY_AGGRO1                      "Ours is the true Horde! The only Horde!"
@@ -218,8 +218,8 @@ struct boss_warchief_kargath_bladefistAI : public ScriptedAI
                     {
                         //move in bladedance
                         float x,y,randx,randy;
-                        randx = (rand()%40);
-                        randy = (rand()%40);
+                        randx = float(rand()%40);
+                        randy = float(rand()%40);
                         x = 210+ randx ;
                         y = -60- randy ;
                         (*me).GetMotionMaster()->MovePoint(1,x,y,me->GetPositionZ());
@@ -252,7 +252,7 @@ struct boss_warchief_kargath_bladefistAI : public ScriptedAI
                 Unit *pTarget = NULL;
                 Creature* Summoned;
 
-                for (int i = 0; i < summoned; i++)
+                for (uint32 i = 0; i < summoned; i++)
                 {
                     switch(rand()%3)
                     {
@@ -270,7 +270,7 @@ struct boss_warchief_kargath_bladefistAI : public ScriptedAI
 
         if (resetcheck_timer <= diff)
         {
-            uint32 tempx,tempy;
+            float tempx,tempy;
             tempx = me->GetPositionX();
             tempy = me->GetPositionY();
             if (tempx > 255 || tempx < 205)
